Created entityTestFactory, to reduce code repitition in entity tests. In process of converting old tests over. Also using uid from res.locals.user to set createdBy fields, instead of client input.
