+++ api/src/controllers/fermentables.js
@@ -1,4 +1,3 @@
-import { isExistingUid } from "./users.js";
 import * as fermentableService from "../services/fermentable.js";
 import * as validate from "../middleware/input-validation.js";
 import { isExistingBreweryUuid } from "./breweries.js";
@@ -8,7 +7,6 @@ import { rejectOnFalse, numberValidator } from "../utils/helpers.js";
 //validation helpers
 
 const opt = { checkFalsy: true };
-const numOpt = { no_symbols: true };
 
 const fermentableUuidChecker = (input) =>
   fermentableService.isExistingFermentableAttribute(input, "fermentableUuid");
@@ -84,7 +82,6 @@ export const getFermentables = [
  * @apiParam {String} breweryUuid The brewery's unique identifier
  * @apiBody {String} name A name for the fermentable
  * @apiBody {String} [fermentableUuid] A unique identifier for the fermentable (a v1 UUID)
- * @apiBody {String} createdBy The uid of the user who created the fermentable
  * @apiBody {String} type One of "Grain", "Sugar", "Extract", "Dry Extract", or "Adjunct"
  * @apiBody {Number} yield Percentage soluable sugar by weight (whole number percentage, i.e. "33" = 33%)
  * @apiBody {Number} color Color contribution in SRM
@@ -112,7 +109,6 @@ const createFermentableValidation = [
     .customSanitizer(validate.xssSanitize),
   customFermentableNameValidator,
   validate.body("fermentableUuid").optional(opt).isUUID(1),
-  validate.body("createdBy").exists(opt).isString().custom(isExistingUid),
   validate.body("type").exists(opt).custom(isValidFermentableType),
   validate.body("yield").exists().custom(isPercentage),
   validate
@@ -153,7 +149,10 @@ const createFermentableController = async (req, res, next) => {
   try {
     const fermentableUuid = await fermentableService.createFermentable(
       req.params.breweryUuid,
-      validate.cleanRequestBody(req, { removeUndefined: true })
+      {
+        ...validate.cleanRequestBody(req, { removeUndefined: true }),
+        createdBy: res.locals.user.uid,
+      }
     );
     return res.locals.sendResponse(res, { uuid: fermentableUuid });
   } catch (error) {
